+++ lib/ex_airtable/table.ex
@@ -102,7 +102,7 @@ defmodule ExAirtable.Table do
       @doc """
       Get all records from your Airtable. See `Service.list/3` for details.
 
-      You may wish to override this function in your table module if you want your manual table listings to pass custom parameters (as a `params: %{}` in `opts`) to the Airtable API. See `ExAirtable.Service.list/2` for details.
+      You may wish to override this function in your table module if you want your manual table listings to pass custom parameters (as a `params: keyword_list` in `opts`) to the Airtable API. See `ExAirtable.Service.list/2` for details.
       """
       def list(opts \\ []) do
         Service.list(table(), opts)
@@ -115,7 +115,7 @@ defmodule ExAirtable.Table do
 
       Typically called automatically by a TableSynchronizer
 
-      You may wish to override this function in your table module if you want your cached table listings to pass custom parameters (as a `params: %{}` in `opts`) to the Airtable API. See `ExAirtable.Service.list/2` for details.
+      You may wish to override this function in your table module if you want your cached table listings to pass custom parameters (as a `params: keyword_list` in `opts`) to the Airtable API. See `ExAirtable.Service.list/2` for details.
       """
       def list_async(opts \\ []) do
         Service.list_async(table(), opts)